make BranchTableInfo an inner static class
